Changed from float to BigDecimal for precision and modified schema for the change. Added check for cart quantity and stock quantity by both backend and HTML.

+++ src/main/java/com/fithub/service/shoppingcart/ShoppingCartService.java
@@ -1,5 +1,7 @@
 package com.fithub.service.shoppingcart;
 
+import java.math.BigDecimal;
+
 import com.fithub.domain.ProductDTO;
 import com.fithub.shoppingcart.ShoppingCart;
 
@@ -25,6 +27,6 @@ public interface ShoppingCartService {
 	 * @return the updated shoppingCart
 	 */
 	ShoppingCart updateShoppingCart(ShoppingCart shoppingCart, ProductDTO productDTO, String shoppingCartOperationType,
-			int productQuantityInCart);
+			BigDecimal productQuantityInCart);
 
 }
